Namespace redis-ha template helpers
When using the `stable/redis-ha` chart with other charts, the defined
template helpers can collide with other charts' helpers if they are not
namespaced. We add a chart namespace to all template helpers and apply
the name change throughout the chart. Also, the persistentVolume name
in redis-master-statefulset.yaml did not match its other references so I
corrected that from name to fullname.